+++ packages/fable-ui/src/models/types/component.ts
@@ -1,11 +1,10 @@
-export type TLayoutComponent = "card" | "h_stack" | "v_stack";
+export type TLayoutComponent = "card" | "form" | "h_stack" | "v_stack";
 
 export type TLeafComponent =
   | "accordion"
   | "button"
   | "checkbox"
   | "datepicker"
-  | "form"
   | "input"
   | "markdown"
   | "pagination"